+++ kernel/include/drivers/time.h
@@ -1,7 +1,4 @@
 #pragma once
-
-#pragma once
-
 void init_task_manager();
 void timer_install();
 void timer_set_frequency();
