+++ kernel/include/sys/system.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void reboot();
+void shutdown();
