+++ kernel/src/drivers/time.c
@@ -25,8 +25,7 @@ void timer_set_frequency(int32_t hz) {
 
 void timer_handler(struct regs *r) {
     //log("tick: %d",timer_ticks);
-    trand(timer_ticks);
-    timer_ticks++;
+    trand(++timer_ticks);
     //if (multi_task) {
        //task_switch();  /* Переключаемся */
     //}
@@ -53,7 +52,6 @@ void sleep(uint16_t delay)
     uint64_t current_ticks = timer_get_ticks();
     while (1)
     {
-        tty_printf("[%d + %d < %d] \n",current_ticks,delay, timer_get_ticks());
         if (current_ticks + delay < timer_get_ticks()){
             break;
         }
