+++ kernel/src/io/ports.c
@@ -201,24 +201,7 @@ int isprint(char c) {
 }
 
 
-void reboot() {
-    log("REBOOT");
 
-    uint8_t good = 0x02;
-
-    while (good & 0x02)
-        good = inb(0x64);
-    outb(0x64, 0xFE);
-    asm volatile("hlt");
-}
-
-
-void shutdown(){
-    log("SHUTDOWN");
-    outw(0xB004, 0x2000);
-    outw(0x604, 0x2000);
-    outw(0x4004, 0x3400);
-}
 
 
 int is_com_port(int port) {
